+++ hydra-node/src/Hydra/API/Server.hs
@@ -231,7 +231,7 @@ mkTimedServerOutputFromStateEvent event =
     StateChanged.HeadIsReadyToFanout{..} -> Just ReadyToFanout{..}
     StateChanged.HeadAborted{headId, utxo} -> Just HeadIsAborted{headId, utxo}
     StateChanged.HeadFannedOut{..} -> Just HeadIsFinalized{..}
-    StateChanged.TransactionAppliedToLocalUTxO{..} -> Just TxValid{headId, transactionId = txId tx, transaction = tx}
+    StateChanged.TransactionAppliedToLocalUTxO{..} -> Just TxValid{headId, transactionId = txId tx}
     StateChanged.TxInvalid{..} -> Just $ TxInvalid{..}
     StateChanged.SnapshotConfirmed{..} -> Just SnapshotConfirmed{..}
     StateChanged.IgnoredHeadInitializing{..} -> Just IgnoredHeadInitializing{..}